+++ packages/utils/tool-utils/src/test/snapshotNormalizer.spec.ts
@@ -4,7 +4,7 @@
  */
 
 import { strict as assert } from "assert";
-import { BlobTreeEntry, TreeTreeEntry } from "@fluidframework/protocol-base";
+import { BlobTreeEntry, TreeTreeEntry } from "@fluidframework/driver-utils";
 import { IBlob, ITree } from "@fluidframework/protocol-definitions";
 import {
 	gcBlobPrefix,
