+++ server/routerlicious/packages/services-client/src/gitManager.ts
@@ -5,6 +5,7 @@
 
 import { assert } from "@fluidframework/common-utils";
 import * as resources from "@fluidframework/gitresources";
+import { buildGitTreeHeirarchy } from "@fluidframework/protocol-base";
 import * as api from "@fluidframework/protocol-definitions";
 import { debug } from "./debug";
 import {
@@ -14,48 +15,6 @@ import {
 	IHistorian,
 } from "./storage";
 import { IWholeFlatSummary, IWholeSummaryPayload, IWriteSummaryResponse } from "./storageContracts";
-
-/**
- * Build a tree hierarchy base on a flat tree
- *
- * @param flatTree - a flat tree
- * @param blobsShaToPathCache - Map with blobs sha as keys and values as path of the blob.
- * @param removeAppTreePrefix - Remove `.app/` from beginning of paths when present
- * @returns the hierarchical tree
- */
-function buildGitTreeHeirarchy(
-	flatTree: resources.ITree,
-	blobsShaToPathCache: Map<string, string> = new Map<string, string>(),
-	removeAppTreePrefix = false,
-): api.ISnapshotTreeEx {
-	const lookup: { [path: string]: api.ISnapshotTreeEx } = {};
-	const root: api.ISnapshotTreeEx = { id: flatTree.sha, blobs: {}, trees: {} };
-	lookup[""] = root;
-
-	for (const entry of flatTree.tree) {
-		const entryPath = removeAppTreePrefix ? entry.path.replace(/^\.app\//, "") : entry.path;
-		const lastIndex = entryPath.lastIndexOf("/");
-		const entryPathDir = entryPath.slice(0, Math.max(0, lastIndex));
-		const entryPathBase = entryPath.slice(lastIndex + 1);
-
-		// The flat output is breadth-first so we can assume we see tree nodes prior to their contents
-		const node = lookup[entryPathDir];
-
-		// Add in either the blob or tree
-		if (entry.type === "tree") {
-			const newTree = { id: entry.sha, blobs: {}, commits: {}, trees: {} };
-			node.trees[decodeURIComponent(entryPathBase)] = newTree;
-			lookup[entryPath] = newTree;
-		} else if (entry.type === "blob") {
-			node.blobs[decodeURIComponent(entryPathBase)] = entry.sha;
-			blobsShaToPathCache.set(entry.sha, `/${entryPath}`);
-		} else {
-			throw new Error("Unknown entry type!!");
-		}
-	}
-
-	return root;
-}
 
 export class GitManager implements IGitManager {
 	private readonly blobCache = new Map<string, resources.IBlob>();
